+++ src/main/java/kr/hhplus/be/server/order/presentation/controller/OrderController.java
@@ -2,14 +2,12 @@ package kr.hhplus.be.server.order.presentation.controller;
 
 import kr.hhplus.be.server.order.application.dto.request.OrderRequest;
 import kr.hhplus.be.server.order.application.dto.response.OrderResponse;
-import kr.hhplus.be.server.order.facade.OrderFacadeService;
-import kr.hhplus.be.server.order.interfaces.api.OrderApiSpec;
+import kr.hhplus.be.server.order.application.facade.OrderFacadeService;
+import kr.hhplus.be.server.order.presentation.interfaces.api.OrderApiSpec;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.math.BigDecimal;
 
 @RestController
 @RequestMapping("/api/v1/order")
