+++ src/main/java/kr/hhplus/be/server/product/application/service/command/ProductCommand.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.product.application.command;
+package kr.hhplus.be.server.product.application.service.command;
 
 public class ProductCommand {
     public record StockDecreaseCommand(
