+++ src/main/java/kr/hhplus/be/server/product/presentation/interfaces/api/ProductApiSpec.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.product.interfaces.api;
+package kr.hhplus.be.server.product.presentation.interfaces.api;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -6,12 +6,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
-import kr.hhplus.be.server.product.application.dto.request.ProductRequest;
 import kr.hhplus.be.server.product.application.dto.response.ProductResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
